backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
ultralytics==8.3.86
easyocr==1.7.2
opencv-python-headless==4.10.0.84
numpy==2.1.1
roboflow==1.1.57
python-dotenv==1.0.1
